+++ metadata.rb
@@ -1,6 +1,6 @@
-maintainer       "Propertyshelf, Inc."
-maintainer_email "thomas@propertyshelf.com"
-license          "Apache 2.0"
-description      "Installs/Configures exim4"
-long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.1.0"
+maintainer        "Inqbus"
+maintainer_email  "thomas.massmann@inqbus.de"
+license           "apache2"
+description       "Installs and configures exim4-light."
+long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
+version           "0.1.0"