+++ recipes/default.rb
@@ -1,8 +1,8 @@
 #
-# Cookbook Name:: exim4
+# Cookbook Name:: exim4-light
 # Recipe:: default
 #
-# Copyright 2012, Propertyshelf, Inc.
+# Copyright 2012, Inqbus
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.